Revert "Split out linux specific stuff to separate files (#55314)"
This reverts commit 005538829080268790de0de947b731c40ee57b81.
