+++ cni/pkg/nodeagent/net_test.go
@@ -31,9 +31,19 @@ import (
 
 	"istio.io/api/annotation"
 	"istio.io/istio/cni/pkg/iptables"
+	istiolog "istio.io/istio/pkg/log"
 	"istio.io/istio/pkg/test/util/assert"
 	"istio.io/istio/tools/istio-iptables/pkg/dependencies"
 )
+
+func setupLogging() {
+	opts := istiolog.DefaultOptions()
+	opts.SetDefaultOutputLevel(istiolog.OverrideScopeName, istiolog.DebugLevel)
+	istiolog.Configure(opts)
+	for _, scope := range istiolog.Scopes() {
+		scope.SetOutputLevel(istiolog.DebugLevel)
+	}
+}
 
 type netTestFixture struct {
 	netServer               *NetServer
