+++ cni/pkg/nodeagent/ztunnelserver_test.go
@@ -105,10 +105,10 @@ func TestZtunnelServerHandleConn(t *testing.T) {
 		}
 		ret := make(chan updateResponse, 1)
 		fdF := int(info.Netns.Fd())
-		req := updateRequest{
-			update: r,
-			fd:     &fdF,
-			resp:   ret,
+		req := UpdateRequest{
+			Update: r,
+			Fd:     &fdF,
+			Resp:   ret,
 		}
 
 		conn.On("SendMsgAndWaitForAck", r, &fdF).Times(1).Return(respData, nil)
@@ -185,10 +185,10 @@ func TestZtunnelServerHandleConnWhenConnDies(t *testing.T) {
 		}
 		ret := make(chan updateResponse, 1)
 		fdF := int(info.Netns.Fd())
-		req := updateRequest{
-			update: r,
-			fd:     &fdF,
-			resp:   ret,
+		req := UpdateRequest{
+			Update: r,
+			Fd:     &fdF,
+			Resp:   ret,
 		}
 
 		conn.On("SendMsgAndWaitForAck", r, &fdF).Times(1).Return(nil, fmt.Errorf("sendmsg: broken pipe"))
