+++ tools/istio-iptables/pkg/capture/run.go
@@ -15,9 +15,12 @@ package capture
 
 import (
 	"fmt"
+	"net"
 	"net/netip"
 	"os"
 	"strings"
+
+	"github.com/vishvananda/netlink"
 
 	"istio.io/istio/pkg/log"
 	"istio.io/istio/tools/istio-iptables/pkg/builder"
@@ -248,6 +251,33 @@ func ignoreExists(err error) error {
 		return nil
 	}
 	return err
+}
+
+// configureIPv6Addresses sets up a new IP address on local interface. This is used as the source IP
+// for inbound traffic to distinguish traffic we want to capture vs traffic we do not. This is needed
+// for IPv6 but not IPv4, as IPv4 defaults to `netmask 255.0.0.0`, which allows binding to addresses
+// in the 127.x.y.z range, while IPv6 defaults to `prefixlen 128` which allows binding only to ::1.
+// Equivalent to `ip -6 addr add "::6/128" dev lo`
+func configureIPv6Addresses(cfg *config.Config) error {
+	if !cfg.EnableIPv6 {
+		return nil
+	}
+	link, err := netlink.LinkByName("lo")
+	if err != nil {
+		return fmt.Errorf("failed to find 'lo' link: %v", err)
+	}
+	// Setup a new IP address on local interface. This is used as the source IP for inbound traffic
+	// to distinguish traffic we want to capture vs traffic we do not.
+	// Equivalent to `ip -6 addr add "::6/128" dev lo`
+	address := &net.IPNet{IP: net.ParseIP("::6"), Mask: net.CIDRMask(128, 128)}
+	addr := &netlink.Addr{IPNet: address}
+
+	err = netlink.AddrAdd(link, addr)
+	if ignoreExists(err) != nil {
+		return fmt.Errorf("failed to add IPv6 inbound address: %v", err)
+	}
+	log.Infof("Added ::6 address")
+	return nil
 }
 
 func (cfg *IptablesConfigurator) Run() error {
