+++ tools/istio-iptables/pkg/capture/run_unspecified.go
@@ -27,11 +27,10 @@ var ErrNotImplemented = errors.New("not implemented")
 
 // configureTProxyRoutes configures ip firewall rules to enable TPROXY support.
 // See https://www.envoyproxy.io/docs/envoy/latest/configuration/http/http_filters/original_src_filter
-// nolint:unused
-func configureTProxyRoutes(_ *config.Config) error {
+func configureTProxyRoutes(cfg *config.Config) error {
 	return ErrNotImplemented
 }
 
-func ConfigureRoutes(_ *config.Config) error {
+func ConfigureRoutes(cfg *config.Config) error {
 	return ErrNotImplemented
 }
